refactor: import 가독성 개선 및 타입 분리
