+++ src/hooks/usePasswordInput.ts
@@ -1,4 +1,5 @@
 import React, { useCallback, useState } from 'react';
+
 const usePasswordInput = <T>(initialState: T) => {
   const [value, setValue] = useState(initialState);
