+++ src/index.tsx
@@ -1,8 +1,7 @@
-import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import './index.css';
 import AppRouter from './router/AppRouter';
+import './index.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(<AppRouter />);
